net: wireless: bcmdhd: fix multiple target RTT failure while associated
RTT measurement failed when device is associated as STA and RTT target
is multiple responder due to incorrect buffer length for WLC_GET_PM
ioctl in the driver.
it results in corruption of variables in rtt_status_info_t structure.

rtt_status_info.pm is changed to int type instead of int8 and use
correct buffer lengh for WLC_GET_PM ioctl.

BUG=28561758

+++ drivers/net/wireless/bcmdhd/dhd_rtt.h
@@ -222,7 +222,7 @@ typedef struct rtt_status_info {
     int8 status;   /* current status for the current entry */
     int8 txchain; /* current device tx chain */
     int8 mpc; /* indicate we change mpc mode */
-    int8 pm; /* to save current value of pm */
+    int pm; /* to save current value of pm */
     int8 pm_restore; /* flag to reset the old value of pm */
     int8 cur_idx; /* current entry to do RTT */
     bool all_cancel; /* cancel all request once we got the cancel requet */
